requirements.txt
numpy
Pillow
scikit-learn==0.21.3
scikit-image